requirements.txt
langchain
qdrant-client
psycopg2-binary
pypdf
python-dotenv
sqlalchemy
fastapi
uvicorn
openai
sentence-transformers
python-multipart
flake8
pytest
